nuevo menu, usuario y log terminados

+++ src/Decoyeso/ProduccionBundle/Controller/SecadorController.php
@@ -101,88 +101,7 @@ class SecadorController extends Controller
     	
     }
     
-    
-   
-    
-    
-    public function estadoSecadorAction($secadores)
-    {
-    	//Si es secador de placas
-    	if ($secador->getTipo() == 0) {
-    		return $this->render('DecoyesoProduccionBundle:Secador:secador_placas.html.twig', array(
-    				'secadores' => $secadores,
-    		));
-    	}
-    	//Si es secador de Molduras
-    	if ($secador->getTipo() == 1) {
-    		return $this->render('DecoyesoProduccionBundle:Secador:secador_molduras.html.twig', array(
-    				'secadores' => $secadores,
-    		));
-    	}
-    	
-    }
-    
 
-    public function disponibilidadSecadoresAction($datos)
-    {
-    	$totalCantidadPlacas = $datos["totalCantidadPlacas"];	
-    	$totalCantidadMolduras = $datos["totalCantidadMolduras"];
-    	
-    	$cantidadLugaresLibresEnSecadoresPlacas = $this->cantidadLugaresLibresEnSecadores(0);
-    	$cantidadLugaresLibresEnSecadoresMolduras = $this->cantidadLugaresLibresEnSecadores(1);
-    	
-    	return $this->render('DecoyesoProduccionBundle:Secador:cuadro_disponibilidad_secadores.html.twig', array(
-    			'totalCantidadPlacas' => $totalCantidadPlacas,
-    			'totalCantidadMolduras' => $totalCantidadMolduras,
-    			'cantidadLugaresLibresEnSecadoresPlacas' => $cantidadLugaresLibresEnSecadoresPlacas,
-    			'cantidadLugaresLibresEnSecadoresMolduras' => $cantidadLugaresLibresEnSecadoresMolduras,
-    	));
-    }
-    
-    
-    public function cantidadLugaresLibresEnSecadores($tipo) {
-    	
-    	$em = $this->getDoctrine()->getEntityManager();
-    	
-    	$cantidad = 0;
-    	
-    	$query = $em->createQuery('SELECT COUNT(lu.id) FROM DecoyesoProduccionBundle:LugarSecador lu
-    			JOIN lu.secador se
-    			WHERE lu.disponible = :lu_disponible AND
-    			se.tipo = :se_tipo
-    			');
-    	$query->setParameters(array(
-    			'lu_disponible' => 0,
-    			'se_tipo' => $tipo,
-    			
-    	));
-    	$cantidad = $query->getSingleScalarResult();
-    	
-    	return $cantidad;
-    }
-    
-    public function cantidadLugaresLibreEnSecadore($idSecador) {
-    
-    	$em = $this->getDoctrine()->getEntityManager();
-    
-    	$cantidad = 0;
-    
-    	$query = $em->createQuery('SELECT COUNT(lu.id) FROM DecoyesoProduccionBundle:LugarSecador lu
-    			JOIN lu.secador se
-    			WHERE se.id = :se_id AND
-    			lu.disponible = :lu_disponible
-    			
-    			');
-    	$query->setParameters(array(
-    			'lu_disponible' => 0,
-    			'se_id' => $idSecador,
-    
-    	));
-    	$cantidad = $query->getSingleScalarResult();
-    
-    	return $cantidad;
-    }
-    
     
 
     /**
@@ -196,7 +115,7 @@ class SecadorController extends Controller
         $entity = $em->getRepository('DecoyesoProduccionBundle:Secador')->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Secador entity.');
+            throw $this->createNotFoundException('ERROR: No se encuentró el secador.');
         }
 
         $deleteForm = $this->createDeleteForm($id);
@@ -250,9 +169,9 @@ class SecadorController extends Controller
             $em->persist($entity);
             $em->flush();
             $log = $this->get('log');
-            $log->create($entity, "Secador Creado");
+            $log->create($entity, "Secador creado");
             
-            $this->get('session')->setFlash('msj_info','El secador se ha creado correctamente');
+            $this->get('session')->setFlash('msj_info','El secador se ha creado correctamente.');
 
             return $this->redirect($this->generateUrl('secador_edit', array('id' => $entity->getId())));
             
@@ -275,7 +194,7 @@ class SecadorController extends Controller
         $entity = $em->getRepository('DecoyesoProduccionBundle:Secador')->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Secador entity.');
+            throw $this->createNotFoundException('ERROR: No se encuentró el secador.');
         }
 
         $editForm = $this->createForm(new SecadorType(), $entity);
@@ -286,6 +205,31 @@ class SecadorController extends Controller
             'edit_form'   => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         ));
+    }
+    
+    public function mostrarEstimacionAction() {
+    	
+    	$em = $this->getDoctrine()->getEntityManager();
+    	
+    	$query = $em->createQuery('SELECT s FROM DecoyesoProduccionBundle:Secador s 
+    			JOIN s.lugaresSecador ls WITH ls.disponible = :ls_disponible
+    			ORDER BY s.nombre ASC
+    			');
+    	
+    	
+    	$query->setParameters(array(
+    			'ls_disponible' => 2,
+    	));
+    	
+    	$secadores = $query->getResult();
+    	
+    	return $this->render('DecoyesoProduccionBundle:Secador:admin_estimacion.html.twig', array(
+    			'secadores'      => $secadores,
+
+    	));
+    	
+    	
+    	
     }
 
     /**
@@ -299,7 +243,7 @@ class SecadorController extends Controller
         $entity = $em->getRepository('DecoyesoProduccionBundle:Secador')->find($id);
 
         if (!$entity) {
-            throw $this->createNotFoundException('Unable to find Secador entity.');
+            throw $this->createNotFoundException('ERROR: No se encuentró el secador.');
         }
 
         $editForm   = $this->createForm(new SecadorType(), $entity);
@@ -310,15 +254,69 @@ class SecadorController extends Controller
         $editForm->bindRequest($request);
 
         if ($editForm->isValid()) {
-            $em->persist($entity);
-            $em->flush();
-            
-            $log = $this->get('log');
-            $log->create($entity, "Secador Actualizado");
-            
-            $this->get('session')->setFlash('msj_info','El Secador se ha actualizado correctamente');
+        	$em->persist($entity);
+        	       	
+        	//Si esta ocupado algun lugar sel secador, no se puede modificar la cantidad de lugares
+        	
+        	//Traigo lugares del secador para ver si estan ocupados
+        	$lugaresSecador = $entity->getLugaresSecador();
+        	
+        	//Si le cambiaron la capacidad
 
-            return $this->redirect($this->generateUrl('secador_edit', array('id' => $id)));
+        	if ($entity->getCapacidad() != count($lugaresSecador)) {
+        		        		
+        		//Me fijo si hay lugares ocupados
+        		$hayOcupado = $this->tieneLugarOcupado($entity);
+        		
+        		foreach($lugaresSecador as $l) {
+        			if ($l->getDisponible()) {
+        				$hayOcupado = true;
+        				break;
+        			}
+        		}
+        		
+        		//si no hay ocupados actualizo la cantidad de lugares disponibles
+        		if (!$hayOcupado) {
+
+        			//borro los lugares que hay
+        			foreach($entity->getLugaresSecador() as $l) {
+        				$em->remove($l);
+        			}
+        			
+        			$entity->setLugarSecador();
+        			$em->persist($entity);
+        			$em->flush();
+        			
+        		
+        			//creo la capacidad nueva
+        			for ($i=1; $i<=$entity->getCapacidad(); $i++):
+	        			$lugarSecador = new LugarSecador();
+	        			$lugarSecador->setNombre("Lugar ".$i);
+	        			$lugarSecador->setDisponible(0);
+	        			$lugarSecador->setSecador($entity);
+	        			$em->persist($lugarSecador);
+	        			$entity->addLugarSecador($lugarSecador);
+
+        			endfor;	
+        			
+        			$em->persist($entity);
+        			$em->flush();
+        			
+        			
+        		}
+        		else {
+        			$this->get('session')->setFlash('msj_info','El secador no se pudo actualizar por que tiene lugares ocupados.');
+        			
+        			return $this->redirect($this->generateUrl('secador_edit', array('id' => $id)));
+        		}
+			}
+        	
+        	$log = $this->get('log');
+        	$log->create($entity, "Secador actualizado");
+        		
+        	$this->get('session')->setFlash('msj_info','El secador se ha actualizado correctamente.');
+        	
+        	return $this->redirect($this->generateUrl('secador_edit', array('id' => $id)));
         }
 
         return $this->render('DecoyesoProduccionBundle:Secador:admin_edit.html.twig', array(
@@ -326,6 +324,26 @@ class SecadorController extends Controller
             'edit_form'   => $editForm->createView(),
             'delete_form' => $deleteForm->createView(),
         ));
+    }
+    
+    public function tieneLugarOcupado($secador) {
+    	
+    	$em = $this->getDoctrine()->getEntityManager();
+    	
+    	//Traigo lugares del secador para ver si estan ocupados
+    	$lugaresSecador = $secador->getLugaresSecador();
+    	
+    	//Me fijo si hay lugares ocupados
+    	$hayOcupado = false;
+    	
+    	foreach($lugaresSecador as $l) {
+    		if ($l->getDisponible()) {
+    			$hayOcupado = true;
+    			break;
+    		}
+    	}
+    	
+    	return $hayOcupado;
     }
 
     /**
@@ -344,13 +362,17 @@ class SecadorController extends Controller
             $entity = $em->getRepository('DecoyesoProduccionBundle:Secador')->find($id);
 
             if (!$entity) {
-                throw $this->createNotFoundException('Unable to find Secador entity.');
+                throw $this->createNotFoundException('ERROR: No se encuentró el secador.');
             }
-
+            
+            if ($hayOcupado = $this->tieneLugarOcupado($entity)) {
+            	$this->get('session')->setFlash('msj_info','El secador no se pudo eliminarse por que tiene lugares ocupados.');
+            	return $this->redirect($this->generateUrl('secador'));
+            }
             
             //LOG
             $log = $this->get('log');
-            $log->create($entity, "Secador Eliminado");
+            $log->create($entity, "Secador eliminado");
             
             $em->remove($entity);
             $em->flush();
@@ -379,4 +401,122 @@ class SecadorController extends Controller
     	));
     
     }
+    
+    public function accionDeleteformAction($id)
+    {
+    	$deleteForm = $this->createDeleteForm($id);
+    
+    	return $this->render('CoobixAdminBundle:Default:accion_delete_form.html.twig', array(
+    			'delete_form' => $deleteForm->createView(),
+    			'url' => $this->generateUrl('secador_delete', array('id' => $id)),
+    			'id'=>$id,
+    			'msj'=>'¿Seguro desea eliminar el secador?'
+    	));
+    
+    }
+    
+    
+    
+    //Funciones
+    
+    public function estadoSecadorAction($secadores)
+    {
+    	//Si es secador de placas
+    	if ($secador->getTipo() == 0) {
+    		return $this->render('DecoyesoProduccionBundle:Secador:secador_placas.html.twig', array(
+    				'secadores' => $secadores,
+    		));
+    	}
+    	//Si es secador de Molduras
+    	if ($secador->getTipo() == 1) {
+    		return $this->render('DecoyesoProduccionBundle:Secador:secador_molduras.html.twig', array(
+    				'secadores' => $secadores,
+    		));
+    	}
+    
+    }
+    
+    
+    public function disponibilidadSecadoresAction($datos)
+    {
+    	$totalCantidadPlacas = $datos["totalCantidadPlacas"];
+    	$totalCantidadMolduras = $datos["totalCantidadMolduras"];
+    
+    	$cantidadLugaresLibresEnSecadoresPlacas = $this->cantidadLugaresLibresEnSecadores(0);
+    	$cantidadLugaresLibresEnSecadoresMolduras = $this->cantidadLugaresLibresEnSecadores(1);
+    
+    	return $this->render('DecoyesoProduccionBundle:Secador:cuadro_disponibilidad_secadores.html.twig', array(
+    			'totalCantidadPlacas' => $totalCantidadPlacas,
+    			'totalCantidadMolduras' => $totalCantidadMolduras,
+    			'cantidadLugaresLibresEnSecadoresPlacas' => $cantidadLugaresLibresEnSecadoresPlacas,
+    			'cantidadLugaresLibresEnSecadoresMolduras' => $cantidadLugaresLibresEnSecadoresMolduras,
+    	));
+    }
+    
+    public function vaciarSecadorAction($id) {
+    	
+    	$em = $this->getDoctrine()->getEntityManager();
+    	$entity = $em->getRepository('DecoyesoProduccionBundle:Secador')->find($id);
+    	
+    	if (!$entity) {
+    		throw $this->createNotFoundException('ERROR: No se encuentró el secador.');
+    	}
+    	
+    	foreach($entity->getLugaresSecador() as $l):
+    	 	$l->setDisponible(0);
+    	endforeach;
+    	$em->persist($l);
+    	$em->flush();
+    	
+    	return $this->redirect($this->generateUrl('secador_estado'));
+    	
+    	
+    	
+    }
+    
+    
+    public function cantidadLugaresLibresEnSecadores($tipo) {
+    
+    	$em = $this->getDoctrine()->getEntityManager();
+    
+    	$cantidad = 0;
+    
+    	$query = $em->createQuery('SELECT COUNT(lu.id) FROM DecoyesoProduccionBundle:LugarSecador lu
+    			JOIN lu.secador se
+    			WHERE lu.disponible = :lu_disponible AND
+    			se.tipo = :se_tipo
+    			');
+    	$query->setParameters(array(
+    			'lu_disponible' => 0,
+    			'se_tipo' => $tipo,
+    
+    	));
+    	$cantidad = $query->getSingleScalarResult();
+    
+    	return $cantidad;
+    }
+    
+    public function cantidadLugaresLibreEnSecadore($idSecador) {
+    
+    	$em = $this->getDoctrine()->getEntityManager();
+    
+    	$cantidad = 0;
+    
+    	$query = $em->createQuery('SELECT COUNT(lu.id) FROM DecoyesoProduccionBundle:LugarSecador lu
+    			JOIN lu.secador se
+    			WHERE se.id = :se_id AND
+    			lu.disponible = :lu_disponible
+    
+    			');
+    	$query->setParameters(array(
+    			'lu_disponible' => 0,
+    			'se_id' => $idSecador,
+    
+    	));
+    	$cantidad = $query->getSingleScalarResult();
+    
+    	return $cantidad;
+    }
+    
+    
 }
